+++ evdev/genecodes_py.py
@@ -15,7 +15,7 @@ print(ecodes.__doc__.strip())
 print('"""')
 
 print()
-print("from typing import Final, Dict, List, Union")
+print("from typing import Final, Dict, Tuple, Union")
 print()
 
 for name, value in ecodes.ecodes.items():
@@ -24,24 +24,24 @@ print()
 
 entries = [
     ("ecodes", "Dict[str, int]", "#: Mapping of names to values."),
-    ("bytype", "Dict[int, Dict[int, Union[str, List[str]]]]", "#: Mapping of event types to other value/name mappings."),
-    ("keys",   "Dict[int, Union[str, List[str]]]", "#: Keys are a combination of all BTN and KEY codes."),
-    ("KEY",    "Dict[int, Union[str, List[str]]]", None),
-    ("ABS",    "Dict[int, Union[str, List[str]]]", None),
-    ("REL",    "Dict[int, Union[str, List[str]]]", None),
-    ("SW",     "Dict[int, Union[str, List[str]]]", None),
-    ("MSC",    "Dict[int, Union[str, List[str]]]", None),
-    ("LED",    "Dict[int, Union[str, List[str]]]", None),
-    ("BTN",    "Dict[int, Union[str, List[str]]]", None),
-    ("REP",    "Dict[int, Union[str, List[str]]]", None),
-    ("SND",    "Dict[int, Union[str, List[str]]]", None),
-    ("ID",     "Dict[int, Union[str, List[str]]]", None),
-    ("EV",     "Dict[int, Union[str, List[str]]]", None),
-    ("BUS",    "Dict[int, Union[str, List[str]]]", None),
-    ("SYN",    "Dict[int, Union[str, List[str]]]", None),
-    ("FF",     "Dict[int, Union[str, List[str]]]", None),
-    ("FF_STATUS",  "Dict[int, Union[str, List[str]]]", None),
-    ("INPUT_PROP", "Dict[int, Union[str, List[str]]]", None)
+    ("bytype", "Dict[int, Dict[int, Union[str, Tuple[str]]]]", "#: Mapping of event types to other value/name mappings."),
+    ("keys",   "Dict[int, Union[str, Tuple[str]]]", "#: Keys are a combination of all BTN and KEY codes."),
+    ("KEY",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("ABS",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("REL",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("SW",     "Dict[int, Union[str, Tuple[str]]]", None),
+    ("MSC",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("LED",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("BTN",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("REP",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("SND",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("ID",     "Dict[int, Union[str, Tuple[str]]]", None),
+    ("EV",     "Dict[int, Union[str, Tuple[str]]]", None),
+    ("BUS",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("SYN",    "Dict[int, Union[str, Tuple[str]]]", None),
+    ("FF",     "Dict[int, Union[str, Tuple[str]]]", None),
+    ("FF_STATUS",  "Dict[int, Union[str, Tuple[str]]]", None),
+    ("INPUT_PROP", "Dict[int, Union[str, Tuple[str]]]", None)
 ]
 
 for key, annotation, doc in entries: